// link2rail/projects/ekp/src/app/empty-wagon/api/generated/model/customer-information.ts
/**
 * Empty wagon API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CustomerInformation { 
    /**
     * Identifier of the customer.
     */
    sgvId: string;
    /**
     * Name of the customer.
     */
    sgvName?: string;
    /**
     * Identifier of the company location.
     */
    partnerId: string;
    /**
     * Name of the company location.
     */
    partnerName?: string;
}