// link2rail/projects/ekp/src/app/empty-wagon/api/generated/model/nhm.ts
/**
 * Empty wagon API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface NHM { 
    /**
     * The NHM code.
     */
    code: string;
    /**
     * The description corresponding to the NHM code.
     */
    description?: string;
}