// link2rail/projects/ekp/src/app/empty-wagon/api/generated/model/transit-railway-undertaking.ts
/**
 * Empty wagon API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface TransitRailwayUndertaking { 
    /**
     * The transit railway undertaking identifier of the company responsible for the load run.
     */
    companyCode: string;
    /**
     * The transit railway undertaking name of the company responsible for the load run.
     */
    companyName?: string;
}